requirements.txt
pandas==0.24.2
python-dateutil
matplotlib
isoweek==1.3.3